% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.combine.R
\name{balance.combine}
\alias{balance.combine}
\title{Combine Two Sub-Plots}
\usage{
balance.combine(balance.partition, balance.distribution, size = "first")
}
\arguments{
\item{balance.partition}{A \code{ggplot} object. The "partition" sub-plot.}

\item{balance.distribution}{A \code{ggplot} object. The "distribution" sub-plot.}

\item{size}{A string. Toggles whether to size final figure based on
"first" (partition) or "last" (distribution) figure provided.}
}
\description{
This function combines the "partition" sub-plot with the
 "distribution" sub-plot, preserving scale.
}
\author{
Thom Quinn
}
