% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.compute.R
\name{balance.fromContrast}
\alias{balance.fromContrast}
\title{Transform Samples with the ilr of a Balance}
\usage{
balance.fromContrast(x, contrast)
}
\arguments{
\item{x}{A matrix with rows as samples (N) and columns as components (D).}

\item{contrast}{A vector. One column of a serial binary partition matrix
with values [-1, 0, 1] describing D components.}
}
\value{
A transformation of samples for the balance provided.
}
\description{
Transform Samples with the ilr of a Balance
}
\examples{
library(balance)
data(iris)
x <- iris[,1:4]
sbp <- sbp.fromPBA(x)
balance.fromContrast(x, sbp[,1])

}
\author{
Thom Quinn
}
