% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.plot.R
\name{balance}
\alias{balance}
\title{Calculate and Visualize Balances}
\usage{
balance(...)
}
\arguments{
\item{...}{Arguments to \code{\link{balance.plot}}.}
}
\description{
This function wraps \code{\link{balance.plot}}.
}
