% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{binary}
\alias{binary}
\title{Convert to binary}
\usage{
binary(x)
}
\arguments{
\item{x}{A number}
}
\value{
The binary version of \code{x} as a vector of 0 and 1
}
\description{
This function turns the input number into a binary sequence.
}
\references{
Crawley, Michael J. The R book. John Wiley & Sons, 2012.
}
\author{
Michael J Crawley
}
