% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-global.R
\docType{data}
\name{expenditures}
\alias{expenditures}
\title{Example Compositional Data}
\format{An object of class \code{data.frame} with 20 rows and 5 columns.}
\usage{
data(expenditures)
}
\description{
Taken from the wonderful \code{robCompositions} package.
}
\keyword{datasets}
