% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlr.R
\name{ldaRule}
\alias{ldaRule}
\title{Compute Pair-wise LDA Rule}
\usage{
ldaRule(x, group)
}
\arguments{
\item{x}{A matrix with rows as samples (N) and columns as components (D).}

\item{group}{A character vector. Group or sub-group membership.
Argument passed to \code{propr::propd}.}
}
\value{
A pair-wise LDA matrix.
}
\description{
This function computes a linear discriminant analysis rule,
 S = (BETWEEN VARIANCE) / (WITHIN VARIANCE), for each feature pair.
 The resultant matrix is clustered to make a discriminative tree.
}
\examples{
\dontrun{
library(balance)
data(iris)
x <- iris[1:100,1:4]
y <- iris[1:100,5]
ldaRule(x, y)
}

}
\author{
Thom Quinn
}
