% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_balance_files.R
\name{make_balance_files}
\alias{make_balance_files}
\title{Create balance datasets from sets of trees}
\usage{
make_balance_files(path_to_data, path_to_output, nLim)
}
\arguments{
\item{path_to_data}{A string, path to where the RDS files containing the trees are}

\item{path_to_output}{A string, path to where the RDS files containing the balance data are (to be saved)}

\item{nLim}{The number of replicate trees to look for per parameter set}
}
\value{
The function returns a bunch of zeros. But it saves as many output files that there were files in the input directory. There is one saved file per parameter combination. Each saved file contains as many rows as there were trees. There are 10 columns.
\itemize{
\item{\code{nTrees}: number of tips of the incipient species tree}
\item{\code{tciTrees}: TCI of the incipient species tree}
\item{\code{mintciTrees}: lowest possible TCI for a tree of this size}
\item{\code{maxtciTrees}: largest possible TCI for a tree of this size}
\item{\code{ntciTrees}: normalized TCI of the incipient-species tree (corrected for the effect of tree size)}
\item{\code{nStrees}: number of tips of the good species tree}
\item{\code{tciStrees}: TCI of the good species tree}
\item{\code{mintciStrees}: lowest possible TCI for a tree of this size}
\item{\code{maxtciStrees}: largest possible TCI for a tree of this size}
\item{\code{ntciTrees}: normalized TCI of the good-species tree (corrected for the effect of tree size)}
}
}
\description{
This function reads lists of phylogenetic trees from a tree folder, calculates the total cophenetic index (TCI, Mir et al. 2013) for each tree, and saves data frames summarizing the calculations.
}
\author{
Raphael Scherrer
}
