% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_tci.R
\name{max_tci}
\alias{max_tci}
\title{Calculate maximum total cophenetic index}
\usage{
max_tci(n)
}
\arguments{
\item{n}{The number of tips in the tree}
}
\value{
An integer, the maximum value of the TCI
}
\description{
This function returns the maximum TCI value (Mir et al. 2013) for a binary tree of size \code{n}.
}
\references{
Mir, A., Rosselló, F., & Rotger, L. a. (2013). A new balance index for phylogenetic trees. Mathematical Biosciences, 241(1), 125–136. https://doi.org/10.1016/j.mbs.2012.10.005
}
\keyword{Cophenetic}
\keyword{Index,}
\keyword{Total}
\keyword{imbalance}
\keyword{tree}
