% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pba.R
\docType{class}
\name{pba-class}
\alias{pba-class}
\alias{predict,pba-method}
\title{A pba model S4 class}
\usage{
\S4method{predict}{pba}(object, newdata)
}
\arguments{
\item{object}{A \code{pba} object.}

\item{newdata}{A matrix.}
}
\description{
A pba model S4 class
}
\section{Methods (by generic)}{
\itemize{
\item \code{predict}: Method to deploy \code{pba} model.
}}

\section{Slots}{

\describe{
\item{\code{data}}{A matrix. The original data.}

\item{\code{sbp}}{A matrix. The SBP matrix.}

\item{\code{pba}}{A matrix. The balances.}

\item{\code{totvar}}{A numeric vector. The total variance per balance.}

\item{\code{subvar}}{A numeric vector. The fractional variance per balance.}
}}

