% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbp.R
\name{sbp.fromRandom}
\alias{sbp.fromRandom}
\title{Build SBP Matrix from Random Tree}
\usage{
sbp.fromRandom(x)
}
\arguments{
\item{x}{A matrix with rows as samples (N) and columns as components (D).}
}
\value{
An SBP matrix.
}
\description{
This function builds an SBP from a random tree.
}
\examples{
library(balance)
data(iris)
x <- iris[,1:4]
sbp.fromRandom(x)

}
\author{
Thom Quinn
}
