% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbp.R
\name{sbp.sort}
\alias{sbp.sort}
\title{Sort SBP Matrix}
\usage{
sbp.sort(sbp)
}
\arguments{
\item{sbp}{An SBP matrix.}
}
\value{
An SBP matrix.
}
\description{
Sort SBP Matrix
}
\examples{
library(balance)
data(iris)
x <- iris[,1:4]
sbp <- sbp.fromPBA(x)
sbp.sort(sbp)

}
\author{
Thom Quinn
}
