% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbp.R
\name{sbp.subset}
\alias{sbp.subset}
\title{Subset SBP Matrix}
\usage{
sbp.subset(sbp, ternary = TRUE, ratios = TRUE)
}
\arguments{
\item{sbp}{An SBP matrix.}

\item{ternary}{A boolean. Toggles whether to return
balances representing three components.}

\item{ratios}{A boolean. Toggles whether to return
balances representing two components.}
}
\value{
An SBP matrix.
}
\description{
Subset SBP Matrix
}
\examples{
library(balance)
data(iris)
x <- iris[,1:4]
sbp <- sbp.fromPBA(x)
sbp.subset(sbp)

}
\author{
Thom Quinn
}
