% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlr.R
\name{ssWithin}
\alias{ssWithin}
\title{Calculate Within-Group Log-ratio Sums of Squares}
\usage{
ssWithin(x, group, ...)
}
\arguments{
\item{x}{A matrix with rows as samples (N) and columns as components (D).}

\item{group}{A character vector. Group or sub-group membership.
Argument passed to \code{propr::propd}.}

\item{...}{Other arguments passed to \code{propr::propd}.}
}
\value{
A VLR matrix.
}
\description{
This function calculates the log-ratio sums of squares
 for all components in a matrix.
}
\author{
Thom Quinn
}
