% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlr.R
\name{vlr}
\alias{vlr}
\title{Calculate Log-ratio Variance}
\usage{
vlr(x, alpha = NA)
}
\arguments{
\item{x}{A matrix with rows as samples (N) and columns as components (D).}

\item{alpha}{A double. Defines a hyper-parameter used
by the Box-Cox transformation to approximate log-ratio
variance in the presence of zeros. Skip with NA.}
}
\value{
A VLR matrix.
}
\description{
This function calculates the log-ratio variance
 for all components in a matrix.
}
\examples{
library(balance)
data(iris)
x <- iris[,1:4]
vlr(x)

}
\author{
Thom Quinn
}
