% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{wide2long}
\alias{wide2long}
\title{Make Long Data from Wide Data}
\usage{
wide2long(wide)
}
\arguments{
\item{wide}{A data set in wide format.}
}
\value{
A data set in long format.
}
\description{
Make Long Data from Wide Data
}
\examples{
library(balance)
data(iris)
x <- iris[,1:4]
sbp <- sbp.fromPBA(x)
wide2long(sbp)

}
\author{
Thom Quinn
}
