% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_balance_files.R
\name{make_balance_files}
\alias{make_balance_files}
\title{Create balance datasets from sets of trees}
\usage{
make_balance_files(path_to_data, path_to_output, nLim)
}
\arguments{
\item{path_to_data}{A string, path to where the RDS files containing the trees are}

\item{path_to_output}{A string, path to where the RDS files containing the balance data are (to be saved)}

\item{nLim}{The number of replicate trees to look for per parameter set}
}
\value{
As many zeros as files in \code{path_to_data} if everything went fine
}
\description{
This function turns the input number into a binary sequence.
}
\author{
Raphael Scherrer
}
