% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlr.R
\name{ssWithin}
\alias{ssWithin}
\title{Calculate Within-Group Log-ratio Sums of Squares}
\usage{
ssWithin(x, group, pairwise = FALSE, ...)
}
\arguments{
\item{x}{A matrix with rows as samples (N) and columns as components (D).}

\item{group}{A character vector. Group or sub-group membership.
Argument passed to \code{propr::propd}.}

\item{pairwise}{A logical. Toggles whether to decompose
sums of squares for each log-ratio.}

\item{...}{Other arguments passed to \code{propr::propd}.}
}
\value{
If \code{pairwise = FALSE}, this function returns
 a vector of the sums of squares for each variable.
 If \code{pairwise = TRUE}, this function returns a matrix
 of the sums of squares for each log-ratio.
}
\description{
This function calculates the within-group sums of squares
 for all variables or all log-ratios. This function
 only supports binary outcomes.
}
\author{
Thom Quinn
}
